Change Default (OSX).sublime-keymap to avoid shadowing important key mappings.
Most of the rest of the key mappings here are `super+shift+?`, so I propose that `super+shift+[bik]` be used to avoid shadowing important default mappings.

`super+k` in particular is a prefix key for many useful text editing functions (e.g. case conversions). `super+b` is the build system trigger key and its not inconceivable that someone would define a custom build system in an rst project.